Create README.md
This is my pet project. Its main idea is to keep tracking on progress with quitting bad habits (smoking, drinking too much coffee, eating sugar) and starting good ones (learning English, training, running, eating healthy food). The idea came to me when I saw a checklist sticked to a fridge and thought that It's not comfortable to depend on papers.

For sure, at the moment this project is far away from being perfect. I am improving it and hope that in the future it can become a useful tool available in the AppStore.